backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pillow==10.3.0
pytesseract==0.3.13
openai-whisper==20231117
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
httpx==0.27.0
aiofiles==23.2.1
